refactor: move TableHelper methods to RepoHelper
