+++ src/main/kotlin/se/uulm/snowballr/backend/repository/CriterionTableRepo.kt
@@ -8,8 +8,6 @@ import se.uulm.snowballr.backend.model.SnowballRException.NotFoundException
 import se.uulm.snowballr.backend.model.dto.Criterion
 import se.uulm.snowballr.backend.model.parseUUID
 import se.uulm.snowballr.backend.table.CriterionTable
-import se.uulm.snowballr.backend.table.getProjectEntityId
-import se.uulm.snowballr.backend.table.getUserEntityId
 import se.uulm.snowballr.backend.table.toCriterion
 import snowballr.CriterionOuterClass
 import java.util.UUID
