+++ src/main/kotlin/se/uulm/snowballr/backend/repository/RepoHelper.kt
@@ -2,6 +2,7 @@ package se.uulm.snowballr.backend.repository
 
 import org.jetbrains.exposed.dao.id.EntityID
 import org.jetbrains.exposed.dao.id.IdTable
+import org.jetbrains.exposed.dao.id.UUIDTable
 import org.jetbrains.exposed.sql.Op
 import org.jetbrains.exposed.sql.ResultRow
 import org.jetbrains.exposed.sql.SqlExpressionBuilder
@@ -12,7 +13,12 @@ import org.jetbrains.exposed.sql.statements.UpdateStatement
 import org.jetbrains.exposed.sql.update
 import se.uulm.snowballr.backend.model.EntityType
 import se.uulm.snowballr.backend.model.SnowballRException.EntityNotPersistedException
+import se.uulm.snowballr.backend.model.SnowballRException.NotFoundException
+import se.uulm.snowballr.backend.table.ProjectTable
+import se.uulm.snowballr.backend.table.UserTable
 import java.util.UUID
+
+// === generic repo helpers === //
 
 /**
  * Returns an entity according to the [where] expression or returns null if the entity couldn't be found.
@@ -42,6 +48,25 @@ fun <Key : Any, T : IdTable<Key>, EntT : Any> T.getEntityOrNull(
  */
 fun <Key : Any, T : IdTable<Key>, EntT : Any> T.getEntityByIdOrNull(id: Key, mapper: (ResultRow) -> EntT): EntT? =
     this.getEntityOrNull(mapper) { this@getEntityByIdOrNull.id eq id }
+
+/**
+ * Returns the entity ID of the entity with the given [id] or throws a [NotFoundException] if no such entity exists.
+ * This can be used to reference the entity in other table rows.
+ *
+ * Example:
+ * Table A stores a reference to this table as `entity_id`. To create a row in table A, we can use this method
+ * to get the [EntityID] and then pass it to the `entity_id` column of table A.
+ *
+ * @param id The ID of the entity as [String].
+ * @param entityType The type of the entity.
+ * @return The ID of the entity as [EntityID].
+ */
+private fun UUIDTable.getEntityId(id: UUID, entityType: EntityType): EntityID<UUID> = this
+    .select(this.id)
+    .where { this@getEntityId.id eq id }
+    .map { it[this.id] }
+    .singleOrNull()
+    ?: throw NotFoundException(entityType, id.toString())
 
 /**
  * Combination of using [insertAndGetId] and fetching the created entity by its ID.
@@ -105,3 +130,21 @@ inline fun <Key : Any, T : IdTable<Key>, EntT : Any> T.updateByIdAndGet(
     entityType: EntityType,
     crossinline body: T.(UpdateStatement) -> Unit,
 ): EntT = this.updateAndGet(mapper, entityType, id.toString(), { this@updateByIdAndGet.id eq id }, body)
+
+// === specific repo helpers === //
+
+/**
+ * Returns the entity ID of the user with the passed [id] or throws a [NotFoundException] if the user doesn't
+ * exist.
+ *
+ * @see getEntityId
+ */
+fun getUserEntityId(id: UUID): EntityID<UUID> = UserTable.getEntityId(id, EntityType.USER)
+
+/**
+ * Returns the entity ID of the project with the passed [id] or throws a [NotFoundException] if the project
+ * doesn't exist.
+ *
+ * @see getEntityId
+ */
+fun getProjectEntityId(id: UUID): EntityID<UUID> = ProjectTable.getEntityId(id, EntityType.PROJECT)
